+++ scenes/metroflip_scene_start.c
@@ -15,6 +15,9 @@ void metroflip_scene_start_on_enter(void* context) {
         submenu, "Rav-Kav", MetroflipSceneRavKav, metroflip_scene_start_submenu_callback, app);
 
     submenu_add_item(
+        submenu, "Navigo", MetroflipSceneNavigo, metroflip_scene_start_submenu_callback, app);
+
+    submenu_add_item(
         submenu,
         "CharlieCard",
         MetroflipSceneCharlieCard,