+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 ## v0.2
 
 - Update Rav-Kav parsing to show more data such as transaction logs
+- Add Navigo parser!
+- Bug fixes